requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.5
botocore==1.23.5
cachetools==4.2.4
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-extra-fields==3.0.2
django-filter==21.1
django-redis==5.0.0
django-storages==1.12.3
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
filelock==3.7.1
fitparse==1.2.0
GDAL==3.4.3
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gpxpy==1.5.0
haversine==2.5.1
hiredis==2.0.0
httplib2==0.20.2
idna==3.3
jmespath==0.10.0
kombu==5.2.1
numpy==1.22.3
oauthlib==3.1.1
openpyxl==3.0.9
Pillow==8.4.0
pipenv==2022.6.7
platformdirs==2.5.2
prompt-toolkit==3.0.22
protobuf==3.19.1
psycopg2==2.9.3
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
PyJWT==2.3.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
utm==0.7.0
uWSGI==2.0.20
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.13.3
